requirements.txt
nxp_imu
pyxl320
simplejson
pyserial
numpy
build_utils
mcp3208
